Added delete function

+++ sort_music.py
@@ -5,8 +5,12 @@ import re
 import shutil
 import argparse
 import yaml
+import glob
+import os
 
 instruments_dictionary = {
+    "voice"   : {"include" : [".*voice.*",".*choir.*",".*solo.*"], "exclude" : []},
+    "audio"   : {"include" : [".*audio.*",".*midi.*"], "exclude" : []},
     "flute"   : {"include" : [".*flute.*"], "exclude" : []},
     "oboe"    : {"include" : [".*oboe.*"], "exclude" : []},
     "clarinet" : {"include" : [".*clarinet.*"], "exclude" : [".*clarinet_bass.*"]},
@@ -17,12 +21,13 @@ instruments_dictionary = {
     "trombone" : {"include" : [".*trombone.*"], "exclude" : []},
     "tuba" : {"include" : [".*tuba.*"], "exclude" : []},
     "timpani" : {"include" : [".*timpani.*"], "exclude" : []},
-    "percussion" : {"include" : [".*percussion.*", ".*triangle.*", ".*glockenspiel.*"], "exclude" : []},
+    "percussion" : {"include" : [".*percussion.*", ".*triangle.*", ".*glockenspiel.*", ".*drum_snare.*", ".*piatti.*"], "exclude" : []},
     "harp" : {"include" : [".*harp.*"], "exclude" : []},
     "piano" : {"include" : [".*piano.*", ".*rhythm.*", ".*synth.*"], "exclude" : []},
     "violin_i" : {"include" : [".*violin_i.*", ".*violin_i_ii.*", ".*violin_1.*", ".*violin_1_2.*"], "exclude" : [".*violin_ii", ".*violin_iii.*"]},
     "violin_ii" : {"include" : [".*violin_i_ii.*", ".*violin_1_2.*", ".*violin_ii.*", ".*violin_2.*"], "exclude" : [".*violin_iii.*"]},
     "violin_iii" : {"include" : [".*violin_iii.*", ".*violin_3.*"], "exclude" : []},
+    "viola" : {"include" : [".*viola.*"], "exclude" : []},
     "cello" : {"include" : [".*cello.*", ".*violoncello.*"], "exclude" : []},
     "c-bass" : {"include" : [".*c-bass.*", ".*contrabass.*", ".*guitar_bass.*"], "exclude" : []}
 }
@@ -70,21 +75,25 @@ class Song:
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description = "Sort music PDF files")
-    parser.add_argument('-f', '--forse_update', action = "store",
+    parser.add_argument('-f', '--forse_update', type=bool, action = "store",
                         dest = "update", required = False,
                         help = "Forse overwrite of existing PDF files")
-    parser.add_argument('-l', '--files_list', action = "store",
-                        dest = "files_list", required = False,
-                        help = "Text file containing list of music pieces to work with")
-    parser.add_argument('-d', '--directories', type=bool, action = "store",
+    parser.add_argument('-i', '--directories', type=bool, action = "store",
                         dest = "create_dirs", required = False,
                         help = "Creates directories under INSTRUMENTS folder")
     parser.add_argument('-m', '--music_list', type=bool, action = "store",
                         dest = "music_list", required = False,
                         help = "Lists all music") 
+    parser.add_argument('-d', '--delete_song', type=bool, action = "store",
+                        dest = "delete_song", required = False,
+                        help = "Delete song(s) from all INSTRUMENT folders")
+    parser.add_argument('-c', '--copy_song', type=bool, action = "store",
+                        dest = "copy_song", required = False,
+                        help = "Copy song(s) to all INSTRUMENTs folders")
 
     args = parser.parse_args()
 
+    print("force update = {}".format(args.update))
 
     try:
         with open('config.yml', 'r') as file:
@@ -107,7 +116,7 @@ if __name__ == '__main__':
 
     def list_music_folders(dir_path):
         ''' Lists all directories found under the provided root directory
-            This will be the name of all songs found under the main music folder
+            This will be the names of all songs found under the main music folder
         '''
         p = Path(dir_path)
         directories_list = [f for f in p.iterdir() if f.is_dir()]
@@ -139,14 +148,43 @@ if __name__ == '__main__':
                 for l in filtered_list:
                     file_path = songObj.path_to_files.joinpath(l).with_suffix('.pdf')
                     to_location = Path(conf['destination_dir']).joinpath(key.upper())
-                    new_file_name = '['+songObj.name+']_'+l+'.pdf'
+                    new_file_name = '['+songObj.name+']_'+l.lower()+'.pdf'
                     to_location = to_location.joinpath(new_file_name)
-                    print(to_location)
-                    shutil.copy2(file_path, to_location)
+                    if not to_location.exists():
+                        print(to_location)
+                        shutil.copy2(file_path, to_location)
+                    else:
+                        print("file {} exists".format(to_location))
+
+    def delete_songs(songObj):    
+        ''' Deletes the requested song(s) from all instruments directories
+            dir_path: path to the instruments directory
+            songObj: list of song objects
+        '''
+        for x in instruments_dictionary.keys():
+            # Get the path to the INSTRUMENT
+            where_location = Path(conf['destination_dir']).joinpath(x.upper())
+            # Wildcard for the file name to get all instances
+            file_name = '*'+songObj.name+'*'
+            # Now we have the list of all files contained the title in question
+            files_list = glob.glob(file_name, root_dir=where_location)
+            # Iterate through the list, construct the absolute path and delete the file
+            for i in files_list:
+                file = where_location.joinpath(i)
+                if os.path.exists(file):
+                    os.remove(file)
+                    print("removed file: {}".format(file))
+                else:
+                    print("file does not exist")
 
     list_of_songs = []
-    for x in conf['hope_orchestra_dir']['song']:
-        list_of_songs.append(Song(x, conf['hope_orchestra_dir']['path']))
+    for x in conf['other_dir']['song']:
+        list_of_songs.append(Song(x, conf['other_dir']['path']))
 
     for s in list_of_songs:
-        copy_files(s)
+        if (args.copy_song):
+            copy_files(s)
+        elif (args.delete_song):
+            delete_songs(s)
+    
+
